+++ apiphany-core/src/main/java/org/apiphany/security/ssl/client/SSLHttpExchangeClientBuilder.java
@@ -112,7 +112,7 @@ public class SSLHttpExchangeClientBuilder extends ExchangeClientBuilder {
 	 */
 	public SSLHttpExchangeClientBuilder protocol(final SSLProtocol protocol) {
 		sslProperties.setProtocol(protocol);
-		return this;
+		return SSLProtocol.DEFAULT != protocol ? sslConfigured() : this;
 	}
 
 	/**